Compilation errors fixed for environments, where qreal is not a double
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@886 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ src/qwt_scale_engine.cpp
@@ -879,7 +879,7 @@ QList<double> QwtLog10ScaleEngine::buildMinorTicks(
             return QList<double>();      // no subticks
 
         // substep factor = 10^substeps
-        const double minFactor = qMax( qPow( 10.0, minStep ), 10.0 );
+        const qreal minFactor = qMax( qPow( 10.0, minStep ), qreal( 10.0 ) );
 
         QList<double> minorTicks;
         for ( int i = 0; i < ( int )majorTicks.count(); i++ )
